runes v1

+++ Assets/Scripts/Player/PlayerHealth.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class PlayerHealth : MonoBehaviour
 {
-    public int health = 100;
+    public float health = 100;
     
     public Action OnHpIsGone;
